+++ polkadot/roadmap/implementers-guide/src/node/utility/pvf-host-and-workers.md
@@ -126,6 +126,17 @@ with untrusted code does not have unnecessary access to the file-system. This
 provides some protection against attackers accessing sensitive data or modifying
 data on the host machine.
 
+*Currently this is only supported on Linux.*
+
+### Restricting networking
+
+We also disable networking on PVF threads by disabling certain syscalls, such as
+the creation of sockets. This prevents attackers from either downloading
+payloads or communicating sensitive data from the validator's machine to the
+outside world.
+
+*Currently this is only supported on Linux.*
+
 ### Clearing env vars
 
 We clear environment variables before handling untrusted code, because why give
